Move types export
