+++ package.json
@@ -1,28 +1,31 @@
 {
   "name": "@strum/colors",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "devDependencies": {
-    "@eslint/js": "9.19.0",
+    "@eslint/js": "9.20.0",
     "@types/culori": "2.1.1",
     "@types/mustache": "4.2.5",
     "bun-plugin-dts": "0.3.0",
-    "bun-types": "1.2.0",
+    "bun-types": "1.2.2",
     "culori": "4.0.1",
-    "eslint": "9.19.0",
+    "eslint": "9.20.0",
     "eslint-config-prettier": "10.0.1",
     "eslint-plugin-unicorn": "56.0.1",
     "husky": "9.1.7",
     "lint-staged": "15.4.3",
     "mustache": "4.2.0",
-    "prettier": "3.4.2",
+    "prettier": "3.5.0",
     "prettier-plugin-sort-json": "4.1.1",
-    "publint": "0.3.2",
-    "syncpack": "13.0.0",
+    "publint": "0.3.4",
+    "syncpack": "13.0.2",
     "tailwindcss": "^3.4.17",
-    "typescript-eslint": "8.21.0",
-    "valibot": "1.0.0-beta.14"
+    "typescript-eslint": "8.23.0",
+    "valibot": "1.0.0-beta.15"
   },
   "exports": {
+    ".": {
+      "types": "./dist/index.d.ts"
+    },
     "./tailwind": {
       "types": "./dist/tailwind/index.d.ts",
       "import": "./dist/tailwind/index.js",
@@ -43,8 +46,7 @@
       "import": "./dist/utils/index.js",
       "require": "./dist/utils/index.cjs",
       "default": "./dist/utils/index.js"
-    },
-    "./types": "./dist/types/index.ts"
+    }
   },
   "files": [
     "dist/*"
